RF: Refactor test suite to more closely resemble structure of library
