--- psychopy/tests/test_all_visual/__init__.py
@@ -1,1 +1,0 @@
-from psychopy import visual
