+++ psychopy/tests/test_experiment/test_compileJS.py
