+++ psychopy/tests/test_experiment/test_component_compile_js.py
