+++ psychopy/tests/test_experiment/test_components.py
