+++ psychopy/tests/test_experiment/test_psyexpCompile.py
