+++ psychopy/tests/test_misc/test_clock.py
