+++ psychopy/tests/test_visual/test_all_stimuli.py
