+++ psychopy/tests/test_visual/test_basevisual.py
