+++ psychopy/tests/test_visual/test_brush.py
