+++ psychopy/tests/test_visual/test_color.py
