+++ psychopy/tests/test_visual/test_contains_overlaps.py
