+++ psychopy/tests/test_visual/test_custommouse.py
