+++ psychopy/tests/test_visual/test_framepacking.py
