+++ psychopy/tests/test_visual/test_gamma.py
