+++ psychopy/tests/test_visual/test_projections_interactive.py
