+++ psychopy/tests/test_visual/test_ratingScale.py
