+++ psychopy/tests/test_visual/test_shape.py
