+++ psychopy/tests/test_visual/test_target.py
