+++ psychopy/tests/test_visual/test_winFlipTiming.py
